Fix updating of sim status while editing sequence. (closes #73)
The sequence editor controller now delegates all simulation updates to the parent scope by emitting 'update.sequence' events when the user edits the sequence. This modifies the global sequence, thus simulations triggered by other sources such as changing bonus stats or the recipe will use the currently edited sequence. When the sequence editor is cancelled, it simply sends another update with the copy of the original sequence which was saved when editing mode was started.