Don't penalize wasted actions explicitly.
With an explicit penalty for wasted actions, sequences which had wasted actions at the end of an otherwise valid sequence were being penalized so heavily that the solver was discarding them very rapidly. When the starting sequence was such a specimen, it essentially meant that the solver started from scratch, as if there was no starting sequence.

Instead, wasted actions are now penalized indirectly by virtue of the fact that they increase the sequence length for no increase in quality fitness, which is now penalized as the final tie-breaker fitness value.