Optimize tooltip caching and display.
$http triggers a digest on each response received, which causes significant lag when loading the app or switching languages because we're loading a couple hundred tooltips.

To resolve this, we roll our own async HTTP requests using XMLHttpRequest and only broadcast one event after all the responses have been received.

On the component side, when the tooltip cache is updated, we rebuild the object used to lookup tooltips for action names in each component. This should speed up digests a little bit since it no longer has to evaluate a function for each tooltip.