Monte-Carlo: Disable automatically moving condition-dependent actions and simulate failure more accurately
Currently, (for example) tricks of the trade will automatially be moved to later in the sequence where it is able to be used.
This is quite unintuitive, and undocumented. This is also bad for sequences using Waste Not, since it means a trick might be moved inside its buff duration, screwing up the durability of the entire sequence.
A more intuitive behaviour is "use the action if it's possible, otherwise skip it", which this commit aims to do.

It turns out that flipping the `overrideOnCondition` flag itself isn't enough, since sometimes actions can not activate (wasted), but the simulator continues to think that counters like Comfort Zone will continue to tick.
Thus a bit of extra "revert the step" logic needs to be added.

This partially addresses bug #174