Fix for issue #370: "Byregot stack efficency is calculated as a buff, not base efficency (quality error)"
Issue caused by applying Byregot's Blessing stack bonus as an additive modifier (like GS and Innovation are) instead of a multiplicative modifier (like touches are). Fix involved changing BB's effect on the quality multiplier to be multiplicative and moving it to occur after GS and Innovation are taken into account.

This fix has the added effect of setting the quality multiplier to zero when BB is used without sufficient IQ stacks even when under the effects of Innovation.